# backend/tests/fixtures/openapi/breaking.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NmClaimItem": {
        "additionalProperties": false,
        "properties": {
          "claim_kind": {
            "maxLength": 40,
            "minLength": 1,
            "title": "Claim Kind",
            "type": "string"
          },
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "id": {
            "exclusiveMinimum": 0.0,
            "title": "Id",
            "type": "integer"
          },
          "node_id": {
            "exclusiveMinimum": 0.0,
            "title": "Node Id",
            "type": "integer"
          },
          "summary": {
            "default": "",
            "title": "Summary",
            "type": "string"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "typed_payload": {
            "additionalProperties": true,
            "title": "Typed Payload",
            "type": "object"
          },
          "uncertainty": {
            "maxLength": 24,
            "minLength": 1,
            "title": "Uncertainty",
            "type": "string"
          },
          "visible_from_chapter": {
            "exclusiveMinimum": 0.0,
            "title": "Visible From Chapter",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "claim_kind",
          "uncertainty",
          "confidence",
          "visible_from_chapter",
          "node_id"
        ],
        "title": "NmClaimItem",
        "type": "object"
      },
      "NmClaimsResponse": {
        "additionalProperties": false,
        "properties": {
          "claims": {
            "items": {
              "$ref": "#/components/schemas/NmClaimItem"
            },
            "title": "Claims",
            "type": "array"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "node_id": {
            "exclusiveMinimum": 0.0,
            "title": "Node Id",
            "type": "integer"
          },
          "novel_id": {
            "exclusiveMinimum": 0.0,
            "title": "Novel Id",
            "type": "integer"
          },
          "publication_status": {
            "const": "candidate_preview",
            "default": "candidate_preview",
            "title": "Publication Status",
            "type": "string"
          },
          "through_chapter": {
            "exclusiveMinimum": 0.0,
            "title": "Through Chapter",
            "type": "integer"
          },
          "version_id": {
            "exclusiveMinimum": 0.0,
            "title": "Version Id",
            "type": "integer"
          }
        },
        "required": [
          "novel_id",
          "version_id",
          "node_id",
          "through_chapter"
        ],
        "title": "NmClaimsResponse",
        "type": "object"
      },
      "NmSourceLinkItem": {
        "additionalProperties": false,
        "properties": {
          "chapter_number": {
            "exclusiveMinimum": 0.0,
            "title": "Chapter Number",
            "type": "integer"
          },
          "claim_id": {
            "exclusiveMinimum": 0.0,
            "title": "Claim Id",
            "type": "integer"
          },
          "content_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Hash"
          },
          "evidence_node_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Evidence Node Id",
            "type": "string"
          },
          "hierarchy_build_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Hierarchy Build Id",
            "type": "string"
          },
          "id": {
            "exclusiveMinimum": 0.0,
            "title": "Id",
            "type": "integer"
          },
          "optional_source_ref": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Optional Source Ref"
          },
          "source_end": {
            "exclusiveMinimum": 0.0,
            "title": "Source End",
            "type": "integer"
          },
          "source_kind": {
            "maxLength": 24,
            "minLength": 1,
            "title": "Source Kind",
            "type": "string"
          },
          "source_start": {
            "minimum": 0.0,
            "title": "Source Start",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "claim_id",
          "source_kind",
          "hierarchy_build_id",
          "evidence_node_id",
          "chapter_number",
          "source_start",
          "source_end"
        ],
        "title": "NmSourceLinkItem",
        "type": "object"
      },
      "NmSourceLinksResponse": {
        "additionalProperties": false,
        "properties": {
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "node_id": {
            "exclusiveMinimum": 0.0,
            "title": "Node Id",
            "type": "integer"
          },
          "novel_id": {
            "exclusiveMinimum": 0.0,
            "title": "Novel Id",
            "type": "integer"
          },
          "publication_status": {
            "const": "candidate_preview",
            "default": "candidate_preview",
            "title": "Publication Status",
            "type": "string"
          },
          "source_links": {
            "items": {
              "$ref": "#/components/schemas/NmSourceLinkItem"
            },
            "title": "Source Links",
            "type": "array"
          },
          "through_chapter": {
            "exclusiveMinimum": 0.0,
            "title": "Through Chapter",
            "type": "integer"
          },
          "version_id": {
            "exclusiveMinimum": 0.0,
            "title": "Version Id",
            "type": "integer"
          }
        },
        "required": [
          "novel_id",
          "version_id",
          "node_id",
          "through_chapter"
        ],
        "title": "NmSourceLinksResponse",
        "type": "object"
      },
      "NmStructureNode": {
        "additionalProperties": false,
        "properties": {
          "chapter_end": {
            "exclusiveMinimum": 0.0,
            "title": "Chapter End",
            "type": "integer"
          },
          "chapter_start": {
            "exclusiveMinimum": 0.0,
            "title": "Chapter Start",
            "type": "integer"
          },
          "child_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Child Ids",
            "type": "array"
          },
          "display_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Label"
          },
          "id": {
            "exclusiveMinimum": 0.0,
            "title": "Id",
            "type": "integer"
          },
          "node_key": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Node Key",
            "type": "string"
          },
          "node_kind": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Node Kind",
            "type": "string"
          }
        },
        "required": [
          "id",
          "node_key",
          "node_kind",
          "chapter_start",
          "chapter_end"
        ],
        "title": "NmStructureNode",
        "type": "object"
      },
      "NmStructureTreeResponse": {
        "additionalProperties": false,
        "properties": {
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/NmStructureNode"
            },
            "title": "Nodes",
            "type": "array"
          },
          "novel_id": {
            "exclusiveMinimum": 0.0,
            "title": "Novel Id",
            "type": "integer"
          },
          "publication_status": {
            "const": "candidate_preview",
            "default": "candidate_preview",
            "title": "Publication Status",
            "type": "string"
          },
          "readiness": {
            "enum": [
              "empty",
              "incomplete",
              "preview_eligible",
              "sealed_candidate"
            ],
            "title": "Readiness",
            "type": "string"
          },
          "through_chapter": {
            "exclusiveMinimum": 0.0,
            "title": "Through Chapter",
            "type": "integer"
          },
          "version_id": {
            "exclusiveMinimum": 0.0,
            "title": "Version Id",
            "type": "integer"
          }
        },
        "required": [
          "novel_id",
          "version_id",
          "through_chapter",
          "readiness"
        ],
        "title": "NmStructureTreeResponse",
        "type": "object"
      },
      "NmVersionListItem": {
        "additionalProperties": false,
        "properties": {
          "badge": {
            "const": "candidate_preview",
            "default": "candidate_preview",
            "title": "Badge",
            "type": "string"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "has_manifest": {
            "default": false,
            "title": "Has Manifest",
            "type": "boolean"
          },
          "node_counts": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "integer"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Node Counts"
          },
          "readiness": {
            "enum": [
              "empty",
              "incomplete",
              "preview_eligible",
              "sealed_candidate"
            ],
            "title": "Readiness",
            "type": "string"
          },
          "validation_verdict": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validation Verdict"
          },
          "version_id": {
            "exclusiveMinimum": 0.0,
            "title": "Version Id",
            "type": "integer"
          },
          "version_key": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Version Key",
            "type": "string"
          }
        },
        "required": [
          "version_id",
          "version_key",
          "readiness"
        ],
        "title": "NmVersionListItem",
        "type": "object"
      },
      "NmVersionListResponse": {
        "additionalProperties": false,
        "properties": {
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "novel_id": {
            "exclusiveMinimum": 0.0,
            "title": "Novel Id",
            "type": "integer"
          },
          "publication_status": {
            "const": "candidate_preview",
            "default": "candidate_preview",
            "title": "Publication Status",
            "type": "string"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/NmVersionListItem"
            },
            "title": "Versions",
            "type": "array"
          }
        },
        "required": [
          "novel_id",
          "versions"
        ],
        "title": "NmVersionListResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "AI 辅助小说创作与理解平台",
    "title": "NovelMind API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/narrative-memory/{novel_id}/versions": {
      "get": {
        "description": "List candidate NM versions for an owned novel (no default active pick).",
        "operationId": "get_versions_api_narrative_memory__novel_id__versions_get",
        "parameters": [
          {
            "in": "path",
            "name": "novel_id",
            "required": true,
            "schema": {
              "title": "Novel Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NmVersionListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Versions",
        "tags": [
          "叙事记忆结构"
        ]
      }
    },
    "/api/narrative-memory/{novel_id}/versions/{version_id}/nodes/{node_id}/claims": {
      "get": {
        "description": "Return claims on a node visible at through_chapter.",
        "operationId": "get_node_claims_api_narrative_memory__novel_id__versions__version_id__nodes__node_id__claims_get",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "title": "Node Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "novel_id",
            "required": true,
            "schema": {
              "title": "Novel Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "through_chapter",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Through Chapter"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NmClaimsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Node Claims",
        "tags": [
          "叙事记忆结构"
        ]
      }
    },
    "/api/narrative-memory/{novel_id}/versions/{version_id}/nodes/{node_id}/source-links": {
      "get": {
        "description": "Return evidence source links for claims on a node (cutoff-filtered).",
        "operationId": "get_node_source_links_api_narrative_memory__novel_id__versions__version_id__nodes__node_id__source_links_get",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "title": "Node Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "novel_id",
            "required": true,
            "schema": {
              "title": "Novel Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "through_chapter",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Through Chapter"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NmSourceLinksResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Node Source Links",
        "tags": [
          "叙事记忆结构"
        ]
      }
    },
    "/api/narrative-memory/{novel_id}/versions/{version_id}/tree": {
      "get": {
        "description": "Return candidate structure tree filtered by through_chapter.",
        "operationId": "get_structure_tree_api_narrative_memory__novel_id__versions__version_id__tree_get",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "novel_id",
            "required": true,
            "schema": {
              "title": "Novel Id",
              "type": "integer"
            }
          },
          {
            "description": "Spoiler cutoff chapter; nodes with chapter_end beyond this are hidden",
            "in": "query",
            "name": "through_chapter",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Spoiler cutoff chapter; nodes with chapter_end beyond this are hidden",
              "title": "Through Chapter"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NmStructureTreeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Structure Tree",
        "tags": [
          "叙事记忆结构"
        ]
      }
    },
...
  }
}